dbusmonitor: reduce number of dbus signals to one per service
Still needs testing & review!

Also line 200 and line 203 cause for a double lookup in the serviceIds
dict. That can probably be done in one lookup.

Summary:
DbusMonitor used to create a signal match for each path it needs to
monitor. Now it just creates one signal for each service. And upon
receiving that signal it looks up the proper item relating to that
path. That looking-up was also done in the old situation be the way.

Changes related to startup time:
- Instead of creating one VeDbusItemImport object for each monitored
  path, we create one signal match per service.

Changes in PropertiesChanged signal handling:
- It uses a new dictionary, self.serviceIds, with the dbus serviceid as
  the key. Everytime a dbus signal comes in, the dictionary is used to
  check if we need to handle that data. This reduces scheduling all kinds
  of tasks in the mainloop for irrelevant changes.
- Also at the same time retrieve and store the 'options' related to that
  path. Which eliminates looking up the 'vrmDict' in the self.items
  dictionary.

Performance improvements:
- startup time reduces from 17 to 5 seconds on a ccgx with 17 solarchargers,
  and from 9 to 3 seconds on a ccgx with vebus, bmv and one solarcharger.
- handling lots of nameownerchange signals has improved. From 15% to 6% CPU
  in the ssh-tunnel test, see [1].
- handling lots of PropertiesChanged signals has not really improved (sadly).
  Tested by running a dummydbusservice generating `lots` of changes, multiple
  every second.

The load caused by the PropertiesChanged signals is not in dbusmonitor or
vrmlogger: putting a simple return statement on line 186 does not reduce the
load. When running the dummydbusservice with 9 relevant paths at a 400ms update
interval, vrmlogger CPU load remains at ~9% both with and without the return
statement.

Putting a simple return statement in connection.py [2], part of the used
dbus-python library, does have a good effect: next place to investigate
is there.

Startup time could be further improved by requesting GetValue on the root
item, instead of doing such a request for every path.

[1] https://github.com/victronenergy/venus/issues/128
[2] https://github.com/mpvader/dbus-python/blob/master/dbus/connection.py#L521