dbusmonitor: reduce number of dbus signals to one per service
Still needs testing & review!

Changes in performance:
- startup time reduces from 17 to 5 seconds on a ccgx with 17 solarchargers,
  and from 9 to 3 seconds on a ccgx with vebus, bmv and one solarcharger.
- handling lots of nameownerchange signals has improved. From 15% to 6% CPU
  in the ssh-tunnel test, see [1].
- handling lots of PropertiesChanged signals has not really improved (sadly).
  Tested by running a dummydbusservice generating `lots` of changes, multiple
  every second.

The load caused by the PropertiesChanged signals is not in dbusmonitor or
vrmlogger: putting a simple return statement on line 186 does not reduce the
load. When running the dummydbusservice with 9 relevant paths at a 400ms update
interval, vrmlogger CPU load remains at ~9% both with and without the return
statement.

Putting a simple return statement in connection.py [2], part of the used
dbus-python library, does have a good effect: next place to investigate
is there.

Startup time could be further improved by requesting GetValue on the root
item, instead of doing such a request for every path.

[1] https://github.com/victronenergy/venus/issues/128
[2] https://github.com/mpvader/dbus-python/blob/master/dbus/connection.py#L521