Refactoring to allow the converter (when used) to access GetText.
Refactor value-tracking into a dedicated object rather than a list.
Add lazy text parameter that calls GetText. Pass this object to
the converter (if supported) so it can access the text
representation should it want to. The text attribute is a lazy
attribute and the text is only fetched if accessed. Startup time
is therefore unaffected, at the expense of a dbus call whenever
the text attribute is accessed. So far, that's limited to
certain firmware versions.